added cauchy loss parameter for radar. Also fixed a bug where we should have been using the extrinsics for the loc sequence.

+++ src/vtr_testing_radar/script/test_localization.sh
@@ -18,17 +18,17 @@ echo "Running localization on sequence ${LOC_INPUT} to reference sequence ${ODO_
 source ${VTRRROOT}/install/setup.bash
 
 graph_dir=${VTRRRESULT}/${ODO_INPUT}/${LOC_INPUT}/graph
-# if [ -d $graph_dir ]; then
-#   # Count the number of directories inside "graph"
-#   dir_count=$(ls -l $graph_dir | grep -c ^d)
+if [ -d $graph_dir ]; then
+  # Count the number of directories inside "graph"
+  dir_count=$(ls -l $graph_dir | grep -c ^d)
   
-#   if [ $dir_count -gt 1 ]; then
-#     read -p "The directory $graph_dir is not empty and contains $dir_count other directories. Do you want to delete it and create an empty one? (yes/no) " response
-#     if [ "$response" == "no" ]; then
-#       exit
-#     fi
-#   fi
-# fi
+  if [ $dir_count -gt 1 ]; then
+    read -p "The directory $graph_dir is not empty and contains $dir_count other directories. Do you want to delete it and create an empty one? (yes/no) " response
+    if [ "$response" == "no" ]; then
+      exit
+    fi
+  fi
+fi
 
 rm -r ${VTRRRESULT}/${ODO_INPUT}/${LOC_INPUT}
 mkdir -p ${VTRRRESULT}/${ODO_INPUT}/${LOC_INPUT}
